message

+++ src/main/java/com/witcher/e_commerce/application/witcher/dao/CategoryRepository.java
@@ -1,11 +1,12 @@
 package com.witcher.e_commerce.application.witcher.dao;
-/*
+
 import com.witcher.e_commerce.application.witcher.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+
 
 }
-*/
